requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
urllib3==2.5.0
